+++ tests/testthat/test-consistency-auditPrior.R
@@ -2,7 +2,7 @@ context("2. Test consistency of function auditPrior()")
 
 # jfa version 0.2.0
 
-test_that(desc = "(id: f4-v0.1.0-t1) Test for method = 'none'", {
+test_that(desc = "(id: f2-v0.1.0-t1) Test for method = 'none'", {
   prior <- auditPrior(confidence = 0.95, likelihood = "binomial", method = "none")
   expect_equal(prior[["description"]]$alpha, 1)
   expect_equal(prior[["description"]]$beta, 1)
@@ -16,7 +16,7 @@ test_that(desc = "(id: f4-v0.1.0-t1) Test for method = 'none'", {
   expect_equal(prior[["description"]]$beta, 1)
 })
 
-test_that(desc = "(id: f4-v0.1.0-t2) Test for method = 'median'", {
+test_that(desc = "(id: f2-v0.1.0-t2) Test for method = 'median'", {
   prior <- auditPrior(confidence = 0.95, likelihood = "binomial", method = "median", materiality = 0.05)
   expect_equal(prior[["description"]]$alpha, 1)
   expect_equal(prior[["description"]]$beta, 13.51341, tolerance = 0.001)
@@ -30,7 +30,7 @@ test_that(desc = "(id: f4-v0.1.0-t2) Test for method = 'median'", {
   expect_equal(prior[["description"]]$beta, 13.86294, tolerance = 0.001)  
 })
 
-test_that(desc = "(id: f4-v0.1.0-t3) Test for method = 'hypotheses'", {
+test_that(desc = "(id: f2-v0.1.0-t3) Test for method = 'hypotheses'", {
   prior <- auditPrior(materiality = 0.05, confidence = 0.95, method = "hypotheses", likelihood = "binomial", pHmin = 0.3)
   expect_equal(prior[["description"]]$alpha, 1)
   expect_equal(prior[["description"]]$beta, 6.954, tolerance = 0.001)
@@ -44,7 +44,7 @@ test_that(desc = "(id: f4-v0.1.0-t3) Test for method = 'hypotheses'", {
   expect_equal(prior[["description"]]$beta, 7.133, tolerance = 0.001) 
 })
 
-test_that(desc = "(id: f4-v0.1.0-t4) Test for method = 'arm'", {
+test_that(desc = "(id: f2-v0.1.0-t4) Test for method = 'arm'", {
   prior <- auditPrior(materiality = 0.05, confidence = 0.95, method = "arm", likelihood = "binomial", ir = 0.6, cr = 0.6)
   expect_equal(prior[["description"]]$alpha, 1)
   expect_equal(prior[["description"]]$beta, 21, tolerance = 0.001)
@@ -58,7 +58,7 @@ test_that(desc = "(id: f4-v0.1.0-t4) Test for method = 'arm'", {
   expect_equal(prior[["description"]]$beta, 21, tolerance = 0.001)
 })
 
-test_that(desc = "(id: f4-v0.1.0-t5) Test for method = 'sample'", {
+test_that(desc = "(id: f2-v0.1.0-t5) Test for method = 'sample'", {
   prior <- auditPrior(materiality = 0.05, confidence = 0.95, method = "sample", likelihood = "binomial", sampleN = 30, sampleK = 1)
   expect_equal(prior[["description"]]$alpha, 2)
   expect_equal(prior[["description"]]$beta, 30)
@@ -72,7 +72,7 @@ test_that(desc = "(id: f4-v0.1.0-t5) Test for method = 'sample'", {
   expect_equal(prior[["description"]]$beta, 30)
 })
 
-test_that(desc = "(id: f4-v0.1.0-t6) Test for method = 'factor'", {
+test_that(desc = "(id: f2-v0.1.0-t6) Test for method = 'factor'", {
   prior <- auditPrior(materiality = 0.05, confidence = 0.95, method = "factor", likelihood = "binomial", sampleN = 30, sampleK = 1, factor = 0.6)
   expect_equal(prior[["description"]]$alpha, 1.6, tolerance = 0.001)
   expect_equal(prior[["description"]]$beta, 18.4, tolerance = 0.001)
@@ -94,7 +94,7 @@ test_that(desc = "(id: f4-v0.1.0-t6) Test for method = 'factor'", {
 
 # jfa version 0.4.0
 
-test_that(desc = "(id: f4-v0.4.0-t1) Test for method = 'median' with expected errors > 0", {
+test_that(desc = "(id: f2-v0.4.0-t1) Test for method = 'median' with expected errors > 0", {
   prior <- auditPrior(materiality = 0.05, confidence = 0.95, method = "median", likelihood = "binomial", expectedError = 0.01)
   expect_equal(prior[["description"]]$alpha, 1.1554, tolerance = 0.001)
   expect_equal(prior[["description"]]$beta, 16.3846, tolerance = 0.001)
@@ -114,14 +114,14 @@ test_that(desc = "(id: f4-v0.4.0-t1) Test for method = 'median' with expected er
 
 # jfa version 0.5.0
 
-test_that(desc = "(id: f4-v0.5.0-t1) Test for print function", {
+test_that(desc = "(id: f2-v0.5.0-t1) Test for print function", {
   prior <- auditPrior(confidence = 0.95, likelihood = "binomial", method = "none")
   invisible(capture.output(print(prior)))
   expect_equal(prior[["description"]]$alpha, 1)
   expect_equal(prior[["description"]]$beta, 1)
 })
 
-test_that(desc = "(id: f4-v0.5.0-t2) Test for plot function", {
+test_that(desc = "(id: f2-v0.5.0-t2) Test for plot function", {
   prior <- auditPrior(confidence = 0.95, likelihood = "binomial", method = "none", materiality = 0.05)
   invisible(capture.output(plot(prior)))
   expect_equal(prior[["description"]]$alpha, 1)
@@ -143,7 +143,7 @@ test_that(desc = "(id: f4-v0.5.0-t2) Test for plot function", {
 
 # jfa version 0.5.3
 
-test_that(desc = "(id: f4-v0.5.3-t1) Test for bram method binomial", {
+test_that(desc = "(id: f2-v0.5.3-t1) Test for bram method binomial", {
   N <- 20000
   materiality <- 2000
   expectedMisstatement <- 300
@@ -156,7 +156,7 @@ test_that(desc = "(id: f4-v0.5.3-t1) Test for bram method binomial", {
   expect_equal(prior[["description"]]$beta, 11.3819)
 })
 
-test_that(desc = "(id: f4-v0.5.3-t2) Test for bram method poisson", {
+test_that(desc = "(id: f2-v0.5.3-t2) Test for bram method poisson", {
   N <- 20000
   materiality <- 2000
   expectedMisstatement <- 300